move distribution.py into distribution package and split into different file for better scalability
