+++ python/setup.py.in
@@ -276,6 +276,7 @@ packages=['paddle',
           'paddle.incubate.tensor',
           'paddle.incubate.nn',
           'paddle.incubate.passes',
+          'paddle.distribution',
           'paddle.distributed.fleet',
           'paddle.distributed.fleet.base',
           'paddle.distributed.fleet.elastic',
